// packages/cp-design-system/src/tokens/atlassian/motion.generated.ts
export const atlassianDurations = {
  instant: 0,
  long: 250,
  medium: 200,
  short: 150,
  xlong: 400,
  xshort: 100,
  xxlong: 600,
  xxshort: 50,
} as const;

export type Bezier = readonly [number, number, number, number];

export const atlassianEasings = {
  'in.practical': [0.6, 0, 0.8, 0.6],
  'inout.bold': [0.4, 0, 0, 1],
  'out.practical': [0.4, 1, 0.6, 1],
  'out.bold': [0, 0.4, 0, 1],
} as const satisfies Record<string, Bezier>;

export const atlassianInteractions = {
  'avatar.enter': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'avatar.exit': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'blanket.enter': {
    duration: 250,
    easing: [0.4, 0, 0, 1],
  },
  'blanket.exit': {
    duration: 200,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'button.hovered': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'button.pressed': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'flag.enter': {
    duration: 250,
    easing: [0, 0.4, 0, 1],
  },
  'flag.exit': {
    duration: 200,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'flag.reposition': {
    duration: 250,
    easing: [0.4, 0, 0, 1],
  },
  'form.message.enter': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'form.message.exit': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  input: {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'input.selection': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'label.enter': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'label.exit': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'listitem.hovered': {
    duration: 50,
    easing: [0.4, 1, 0.6, 1],
  },
  'listitem.pressed': {
    duration: 100,
    easing: [0.4, 1, 0.6, 1],
  },
  'listitem.selected': {
    duration: 100,
    easing: [0.4, 1, 0.6, 1],
  },
  'modal.enter': {
    duration: 250,
    easing: [0.4, 0, 0, 1],
  },
  'modal.exit': {
    duration: 200,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'panel.content.enter': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'panel.content.exit': {
    duration: 50,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'panel.enter': {
    duration: 250,
    easing: [0, 0.4, 0, 1],
  },
  'panel.enter.left': {
    duration: 250,
    easing: [0, 0.4, 0, 1],
  },
  'panel.enter.right': {
    duration: 250,
    easing: [0, 0.4, 0, 1],
  },
  'panel.exit': {
    duration: 200,
    easing: [0, 0.4, 0, 1],
  },
  'panel.exit.left': {
    duration: 200,
    easing: [0, 0.4, 0, 1],
  },
  'panel.exit.right': {
    duration: 200,
    easing: [0, 0.4, 0, 1],
  },
  'popup.enter.bottom': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'popup.enter.left': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'popup.enter.right': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'popup.enter.top': {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'popup.exit.bottom': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'popup.exit.left': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'popup.exit.right': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'popup.exit.top': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'sidenav.enter.left': {
    duration: 250,
    easing: [0, 0.4, 0, 1],
  },
  'sidenav.enter.right': {
    duration: 250,
    easing: [0, 0.4, 0, 1],
  },
  'sidenav.exit.left': {
    duration: 200,
    easing: [0, 0.4, 0, 1],
  },
  'sidenav.exit.right': {
    duration: 200,
    easing: [0, 0.4, 0, 1],
  },
  'spotlight.enter': {
    duration: 250,
    easing: [0.4, 0, 0, 1],
  },
  'spotlight.exit': {
    duration: 200,
    easing: [0.6, 0, 0.8, 0.6],
  },
  tab: {
    duration: 150,
    easing: [0.4, 1, 0.6, 1],
  },
  'tab.indicator.enter.left': {
    duration: 150,
    easing: [0, 0.4, 0, 1],
  },
  'tab.indicator.enter.right': {
    duration: 150,
    easing: [0, 0.4, 0, 1],
  },
  'tab.indicator.exit.left': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
  'tab.indicator.exit.right': {
    duration: 100,
    easing: [0.6, 0, 0.8, 0.6],
  },
} as const;